// bitirme_projeleri/angular/Buğra Furkan Işınak/src/app/models/IProducts.ts
export interface IProducts {
    meta: Meta;
    data: Product[];
}

export interface ISingleProduct {
    meta: Meta,
    data: Product
}
export interface Lesson {
  id: number;
  lesson_name: string;
  lesson_content: string;
}
export interface Product {
    id:                   number;
    title:                string;
    description:          string;
    category:             string;
    price:                number;
    discountPercentage:   number;
    rating:               number;
    stock:                number;
    tags:                 string[];
    brand:                string;
    sku:                  string;
    minimumOrderQuantity: number;
    images:               string[];
    lessons:               Lesson[];
}


export interface Meta {
    status:     number;
    message:    string;
    pagination: Pagination;
}

export interface Pagination {
    page:        number;
    per_page:    number;
    total_items: number;
    total_pages: number;
}